requirements.txt
dingtalk-stream
dashscope
python-dotenv
Pillow
